+++ check list.txt
@@ -1,4 +1,4 @@
 4th submission
 -- check utils.js updated and fixed
 -- index.html updated adde modal__button_disabled class (line 85,121)
--- Card.js 404 -- testing
+-- Card.js 404 -- testing -- fixed in parrt III - Submission4
